// src/app/pages/program-administration/models/review-document-list.model.ts
export interface ReviewDocumentList {
  docListId: number;
  recType?: string;
  offCd?: string;
  grantPgmId?: number;
  seqNo: number;
  subRvwId?: number;
  docCat?: string;
  docCd?: string;
  docName?: string;
  docType?: string;
  valInd?: string;
  valDt?: string;
  valUnits?: number;
  valUm?: string;
  minNos?: number;
  maxNos?: number;
  fileTypeAppl?: string;
  helpTxt?: string;
  fileTitle?: string;
  attachName?: string;
  createDt?: string;
  createId?: number;
  lastUpdDt?: string;
  lastUpdId?: number;
  subCat?: string;
  rvwType?: string;
  entType?: string;
  formId?: number;
  requireDocumentComment?: boolean;
  requireCheckListComment?: boolean;
  inactive?: boolean;
  defaultDoc?: string;
  documentDueDateDaysOffsetFromReviewStartDate?: number;
  flexFormPdftemplateId?: number;
  allowMultipleFlexformInstances?: boolean;
  unlimitedRecipients?: boolean;
  dataSrc?: string;
  approvalEmailTemplateId?: number;
  editableEmail?: boolean;
  approvalHtml?: string;
  approvalConfirmationEmailTemplateId?: number;
  questionnaireId?: number;
  allowMultipleQuestionnaireInstances?: boolean;
  questionnaireFieldId?: number;
  mapQuestionnaireWorkflowCriteria?: boolean;
  remapQuestionnaire?: boolean;
}